Adding a task to test a "live" noid server
I used the following to verify that the client works against a running
noids instance (instead of mocks and stubs).

When I used the task against the master branch that did not include
updates to rest_client, I got similar errors as the spec suite (e.g.
`KeyError: key not found: :ciphers`)